Move property position for better understanding
