+++ osu.Framework.Font/Graphics/Sprites/LyricSpriteText.cs
@@ -58,6 +58,8 @@ namespace osu.Framework.Graphics.Sprites
             }
         }
 
+        #region text
+
         private string text = string.Empty;
 
         /// <summary>
@@ -87,6 +89,52 @@ namespace osu.Framework.Graphics.Sprites
 
         private string displayedText => Text;
 
+        private PositionText[] rubies;
+
+        /// <summary>
+        /// Gets or sets the ruby text to be displayed.
+        /// </summary>
+        public PositionText[] Rubies
+        {
+            get => rubies;
+            set
+            {
+                rubies = filterValidValues(value);
+
+                invalidate(true);
+                Invalidate(Invalidation.All);
+            }
+        }
+
+        private PositionText[] romajies;
+
+        /// <summary>
+        /// Gets or sets the romaji text to be displayed.
+        /// </summary>
+        public PositionText[] Romajies
+        {
+            get => romajies;
+            set
+            {
+                romajies = filterValidValues(value);
+
+                invalidate(true);
+                Invalidate(Invalidation.All);
+            }
+        }
+
+        private PositionText[] filterValidValues(PositionText[] texts)
+        {
+            string text = Text;
+            return texts?.Where(positionText => Math.Min(positionText.StartIndex, positionText.EndIndex) >= 0
+                                                && Math.Max(positionText.StartIndex, positionText.EndIndex) <= text.Length
+                                                && positionText.EndIndex > positionText.StartIndex).ToArray();
+        }
+
+        #endregion
+
+        #region font
+
         private FontUsage font = FontUsage.Default;
 
         /// <summary>
@@ -103,6 +151,44 @@ namespace osu.Framework.Graphics.Sprites
                 shadowOffsetCache.Invalidate();
             }
         }
+
+        private FontUsage rubyFont = FontUsage.Default;
+
+        /// <summary>
+        /// Contains information on the font used to display the text.
+        /// </summary>
+        public FontUsage RubyFont
+        {
+            get => rubyFont;
+            set
+            {
+                rubyFont = value;
+
+                invalidate(true);
+                shadowOffsetCache.Invalidate();
+            }
+        }
+
+        private FontUsage romajiFont = FontUsage.Default;
+
+        /// <summary>
+        /// Contains information on the font used to display the text.
+        /// </summary>
+        public FontUsage RomajiFont
+        {
+            get => romajiFont;
+            set
+            {
+                romajiFont = value;
+
+                invalidate(true);
+                shadowOffsetCache.Invalidate();
+            }
+        }
+
+        #endregion
+
+        #region style
 
         private bool allowMultiline = true;
 
@@ -209,6 +295,128 @@ namespace osu.Framework.Graphics.Sprites
             }
         }
 
+        private ILyricTexture textTexture;
+
+        public ILyricTexture TextTexture
+        {
+            get => textTexture;
+            set
+            {
+                if (textTexture == value)
+                    return;
+
+                textTexture = value;
+                Colour = (textTexture as SolidTexture)?.SolidColor ?? Color4.White;
+
+                Invalidate(Invalidation.All);
+            }
+        }
+
+        private ILyricTexture shadowTexture;
+
+        public ILyricTexture ShadowTexture
+        {
+            get => shadowTexture;
+            set
+            {
+                if (shadowTexture == value)
+                    return;
+
+                shadowTexture = value;
+                ShadowColour = (shadowTexture as SolidTexture)?.SolidColor ?? Color4.White;
+                Invalidate(Invalidation.All);
+            }
+        }
+
+        private ILyricTexture borderTexture;
+
+        public ILyricTexture BorderTexture
+        {
+            get => borderTexture;
+            set
+            {
+                if (borderTexture == value)
+                    return;
+
+                borderTexture = value;
+                Invalidate(Invalidation.All);
+            }
+        }
+
+        private LyricTextAlignment rubyAlignment;
+
+        /// <summary>
+        /// Gets or sets the ruby alignment.
+        /// </summary>
+        public LyricTextAlignment RubyAlignment
+        {
+            get => rubyAlignment;
+            set
+            {
+                if (rubyAlignment == value)
+                    return;
+
+                rubyAlignment = value;
+                invalidate(true);
+            }
+        }
+
+        private LyricTextAlignment romajiAlignment;
+
+        /// <summary>
+        /// Gets or sets the romaji alignment.
+        /// </summary>
+        public LyricTextAlignment RomajiAlignment
+        {
+            get => romajiAlignment;
+            set
+            {
+                if (romajiAlignment == value)
+                    return;
+
+                romajiAlignment = value;
+                invalidate(true);
+            }
+        }
+
+        private float borderRadius;
+
+        /// <summary>
+        /// Gets or sets the border redius
+        /// </summary>
+        public float BorderRadius
+        {
+            get => borderRadius;
+            set
+            {
+                if (borderRadius == value)
+                    return;
+
+                borderRadius = value;
+
+                invalidate(true);
+            }
+        }
+
+        private bool border;
+
+        /// <summary>
+        /// Gets or sets the border
+        /// </summary>
+        public bool Border
+        {
+            get => border;
+            set
+            {
+                if (border == value)
+                    return;
+
+                border = value;
+
+                invalidate(true);
+            }
+        }
+
         private bool truncate;
 
         /// <summary>
@@ -250,6 +458,10 @@ namespace osu.Framework.Graphics.Sprites
                 invalidate(true);
             }
         }
+
+        #endregion
+
+        #region size
 
         private bool requiresAutoSizedWidth => explicitWidth == null && (RelativeSizeAxes & Axes.X) == 0;
 
@@ -344,6 +556,10 @@ namespace osu.Framework.Graphics.Sprites
             }
         }
 
+        #endregion
+
+        #region text spacing
+
         private Vector2 spacing;
 
         /// <summary>
@@ -362,6 +578,48 @@ namespace osu.Framework.Graphics.Sprites
                 invalidate(true);
             }
         }
+
+        private Vector2 rubySpacing;
+
+        /// <summary>
+        /// Gets or sets the spacing between characters of ruby text.
+        /// </summary>
+        public Vector2 RubySpacing
+        {
+            get => rubySpacing;
+            set
+            {
+                if (rubySpacing == value)
+                    return;
+
+                rubySpacing = value;
+
+                invalidate(true);
+            }
+        }
+
+        private Vector2 romajiSpacing;
+
+        /// <summary>
+        /// Gets or sets the spacing between characters of romaji text.
+        /// </summary>
+        public Vector2 RomajiSpacing
+        {
+            get => romajiSpacing;
+            set
+            {
+                if (romajiSpacing == value)
+                    return;
+
+                romajiSpacing = value;
+
+                invalidate(true);
+            }
+        }
+
+        #endregion
+
+        #region margin/padding
 
         private MarginPadding padding;
 
@@ -384,167 +642,47 @@ namespace osu.Framework.Graphics.Sprites
             }
         }
 
-        public override bool IsPresent => base.IsPresent && (AlwaysPresent || !string.IsNullOrEmpty(displayedText));
-
-        #region Characters
-
-        private readonly LayoutValue charactersCache = new LayoutValue(Invalidation.DrawSize | Invalidation.Presence, InvalidationSource.Parent);
+        private int rubyMargin = 5;
 
         /// <summary>
-        /// Glyph list to be passed to <see cref="TextBuilder"/>.
+        /// Shrinks the space between ruby and main text.
         /// </summary>
-        private readonly List<TextBuilderGlyph> charactersBacking = new List<TextBuilderGlyph>();
-
-        /// <summary>
-        /// The characters in local space.
-        /// </summary>
-        private List<TextBuilderGlyph> characters
+        public int RubyMargin
         {
-            get
+            get => rubyMargin;
+            set
             {
-                computeCharacters();
-                return charactersBacking;
-            }
-        }
-
-        private bool isComputingCharacters;
-
-        /// <summary>
-        /// Compute character textures and positions.
-        /// </summary>
-        private void computeCharacters()
-        {
-            // Note : this feature can only use in osu-framework
-            // if (LoadState >= LoadState.Loaded)
-            //     ThreadSafety.EnsureUpdateThread();
-
-            if (store == null)
-                return;
-
-            if (charactersCache.IsValid)
-                return;
-
-            charactersBacking.Clear();
-
-            Debug.Assert(!isComputingCharacters, "Cyclic invocation of computeCharacters()!");
-            isComputingCharacters = true;
-
-            TextBuilder textBuilder = null;
-
-            try
-            {
-                if (string.IsNullOrEmpty(displayedText))
+                if (rubyMargin == value)
                     return;
 
-                textBuilder = CreateTextBuilder(store);
-                textBuilder.AddText(displayedText);
+                rubyMargin = value;
 
-                // Get main text position and main text array.
-                var existCharacter = textBuilder.Characters.ToArray();
-
-                // Print ruby texts
-                var rubyYPosition = Padding.Top - RubyMargin;
-                createPositionTexts(existCharacter, Rubies, rubyYPosition, true);
-
-                // Calculate position and print romaji texts
-                var textHeight = existCharacter.FirstOrDefault().Height + existCharacter.FirstOrDefault().YOffset;
-                var romajiYPosition = textHeight + RomajiMargin;
-                createPositionTexts(existCharacter, Romajies, romajiYPosition, false);
-            }
-            finally
-            {
-                if (requiresAutoSizedWidth)
-                    base.Width = (textBuilder?.Bounds.X ?? 0) + Padding.Right;
-                if (requiresAutoSizedHeight)
-                    base.Height = (textBuilder?.Bounds.Y ?? 0) + Padding.Bottom;
-
-                base.Width = Math.Min(base.Width, MaxWidth);
-
-                isComputingCharacters = false;
-                charactersCache.Validate();
-            }
-
-            // Create ruby and romaji texts
-            void createPositionTexts(TextBuilderGlyph[] mainTexts, PositionText[] positionTexts, float yPosition, bool ruby)
-            {
-                positionTexts?.ForEach(p =>
-                {
-                    var text = p.Text;
-                    if (string.IsNullOrEmpty(text))
-                        return;
-
-                    // Get text position
-                    var spacing = ruby ? rubySpacing : romajiSpacing;
-                    var textPosition = new Vector2(getTextPosition(p, spacing.X), yPosition);
-
-                    // create ruby or romaji builder
-                    var builder = ruby ? CreateRubyTextBuilder(store, textPosition) : CreateRomajiTextBuilder(store, textPosition);
-                    builder.AddText(text);
-                });
-
-                // Convert
-                float getTextPosition(PositionText text, float textSpacing)
-                {
-                    //It's magic number to let text in the center
-                    const float size_multiple = 1.4f;
-                    var centerPosition = (mainTexts[text.StartIndex].DrawRectangle.Left + mainTexts[text.EndIndex - 1].DrawRectangle.Right) / 2;
-                    var textWidth = text.Text.Sum(c => (store.Get(font.FontName, c)?.Width ?? 0) * font.Size * size_multiple) - (text.Text.Length) * textSpacing;
-                    return centerPosition - textWidth / 2;
-                }
+                invalidate(true);
             }
         }
 
-        private readonly LayoutValue parentScreenSpaceCache = new LayoutValue(Invalidation.DrawSize | Invalidation.Presence | Invalidation.DrawInfo, InvalidationSource.Parent);
-        private readonly LayoutValue localScreenSpaceCache = new LayoutValue(Invalidation.MiscGeometry, InvalidationSource.Self);
-
-        private readonly List<ScreenSpaceCharacterPart> screenSpaceCharactersBacking = new List<ScreenSpaceCharacterPart>();
+        private int romajiMargin;
 
         /// <summary>
-        /// The characters in screen space. These are ready to be drawn.
+        /// Shrinks the space between romaji and main text.
         /// </summary>
-        private List<ScreenSpaceCharacterPart> screenSpaceCharacters
+        public int RomajiMargin
         {
-            get
+            get => romajiMargin;
+            set
             {
-                computeScreenSpaceCharacters();
-                return screenSpaceCharactersBacking;
+                if (romajiMargin == value)
+                    return;
+
+                romajiMargin = value;
+
+                invalidate(true);
             }
         }
-
-        private void computeScreenSpaceCharacters()
-        {
-            if (!parentScreenSpaceCache.IsValid)
-            {
-                localScreenSpaceCache.Invalidate();
-                parentScreenSpaceCache.Validate();
-            }
-
-            if (localScreenSpaceCache.IsValid)
-                return;
-
-            screenSpaceCharactersBacking.Clear();
-
-            Vector2 inflationAmount = DrawInfo.MatrixInverse.ExtractScale().Xy;
-
-            foreach (var character in characters)
-            {
-                screenSpaceCharactersBacking.Add(new ScreenSpaceCharacterPart
-                {
-                    DrawQuad = ToScreenSpace(character.DrawRectangle.Inflate(inflationAmount)),
-                    InflationPercentage = Vector2.Divide(inflationAmount, character.DrawRectangle.Size),
-                    Texture = character.Texture
-                });
-            }
-
-            localScreenSpaceCache.Validate();
-        }
-
-        private readonly LayoutValue<Vector2> shadowOffsetCache = new LayoutValue<Vector2>(Invalidation.DrawInfo, InvalidationSource.Parent);
-
-        private Vector2 premultipliedShadowOffset =>
-            shadowOffsetCache.IsValid ? shadowOffsetCache.Value : shadowOffsetCache.Value = ToScreenSpace(shadowOffset * Font.Size) - ToScreenSpace(Vector2.Zero);
 
         #endregion
+
+        public override bool IsPresent => base.IsPresent && (AlwaysPresent || !string.IsNullOrEmpty(displayedText));
 
         #region Invalidation
 
@@ -557,12 +695,6 @@ namespace osu.Framework.Graphics.Sprites
 
             Invalidate(Invalidation.DrawNode);
         }
-
-        #endregion
-
-        #region DrawNode
-
-        protected override DrawNode CreateDrawNode() => new LyricSpriteTextDrawNode(this);
 
         #endregion
